+++ internal/provider/provider.go
@@ -9,6 +9,7 @@ import (
 	"os"
 
 	tfe "github.com/hashicorp/go-tfe"
+	tfev2 "github.com/hashicorp/go-tfe/v2"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-provider-tfe/internal/client"
@@ -25,6 +26,7 @@ var (
 // specify one.
 type ConfiguredClient struct {
 	Client       *tfe.Client
+	ClientV2     *tfev2.Client
 	Organization string
 }
 
@@ -197,6 +199,7 @@ func configure() schema.ConfigureContextFunc {
 
 		return ConfiguredClient{
 			providerClient.TfeClient,
+			providerClient.TfeClientV2,
 			providerOrganization,
 		}, diagnosticWarnings
 	}